Fixed a few compile issues.

+++ geometry/prismesh.c
@@ -610,7 +610,7 @@ static int64_t* source_vector(prismesh_t* mesh)
     proc_offsets[p+1] = proc_offsets[p] + num_chunks_for_proc[p];
 
   // Gather the indices of the chunks owned by all processes into a huge list.
-  int num_all_chunks = mesh->num_xy_chunks * mesh->num_z_chunks;
+  int num_all_chunks = (int)(mesh->num_xy_chunks * mesh->num_z_chunks);
   ASSERT(num_all_chunks == proc_offsets[mesh->nproc]);
   int* all_chunks = polymec_malloc(sizeof(int) * num_all_chunks);
   MPI_Allgatherv(my_chunks->data, num_my_chunks, MPI_INT, 
@@ -657,10 +657,10 @@ static void redistribute_prismesh_field(prismesh_field_t** field,
   }
 
   // Post receives for each chunk in the new field.
-  int num_new_local_chunks = prismesh_field_num_chunks(new_field);
+  size_t num_new_local_chunks = prismesh_field_num_chunks(new_field);
   MPI_Request recv_requests[num_new_local_chunks];
   pos = 0;
-  int num_recv_reqs = 0;
+  size_t num_recv_reqs = 0;
   while (prismesh_field_next_chunk(new_field, &pos, &xy_index, &z_index, &chunk, &data))
   {
     int ch = chunk_index(new_mesh, xy_index, z_index);
@@ -677,10 +677,10 @@ static void redistribute_prismesh_field(prismesh_field_t** field,
   ASSERT(num_recv_reqs <= num_new_local_chunks);
 
   // Post sends.
-  int num_old_local_chunks = prismesh_field_num_chunks(old_field);
+  size_t num_old_local_chunks = prismesh_field_num_chunks(old_field);
   MPI_Request send_requests[num_old_local_chunks];
   pos = 0;
-  int num_send_reqs = 0;
+  size_t num_send_reqs = 0;
   while (prismesh_field_next_chunk(old_field, &pos, &xy_index, &z_index, &chunk, &data))
   {
     int ch = chunk_index(new_mesh, xy_index, z_index);
@@ -697,8 +697,8 @@ static void redistribute_prismesh_field(prismesh_field_t** field,
   ASSERT(num_send_reqs <= num_old_local_chunks);
 
   // Wait for everything to finish.
-  MPI_Waitall(num_send_reqs, send_requests, MPI_STATUSES_IGNORE);
-  MPI_Waitall(num_recv_reqs, recv_requests, MPI_STATUSES_IGNORE);
+  MPI_Waitall((int)num_send_reqs, send_requests, MPI_STATUSES_IGNORE);
+  MPI_Waitall((int)num_recv_reqs, recv_requests, MPI_STATUSES_IGNORE);
 
   // Replace the old field with the new one.
   *field = new_field;
